+++ spec/support/servers/plugins/webhook_notify.rb
@@ -1,0 +1,9 @@
+require 'sinatra'
+require 'json'
+
+post '/hook' do
+    [
+        request.env['CONTENT_TYPE'],
+        request.body.read
+    ].to_json
+end